AddingBackGround
