+++ package.json
@@ -40,7 +40,8 @@
     "react-native-webview": "^11.2.3",
     "react-navigation": "^4.4.3",
     "react-navigation-stack": "^2.10.2",
-    "redux-thunk": "^2.3.0"
+    "redux-thunk": "^2.3.0",
+    "expo-asset": "~8.2.0"
   },
   "devDependencies": {
     "@babel/core": "~7.9.0",
