+++ screens/MakeAssistanceRequestScreen.js
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from 'react';
-import { StyleSheet, Text, View, TextInput, TouchableOpacity} from 'react-native';
+import { StyleSheet, Text, View, TextInput, TouchableOpacity, ImageBackground} from 'react-native';
 import {Card} from 'react-native-paper';
 import * as firebase from 'firebase';
 import * as Location from 'expo-location';
@@ -70,7 +70,7 @@ const MakeAssistanceRequestScreen = props => {
 
 
     return(
-
+<ImageBackground source={require('../assets/BG.png')} style={styles.bImage}>
       <KeyboardAwareScrollView style={styles.container}>
         <View style={styles.screen} >
 
@@ -125,6 +125,7 @@ const MakeAssistanceRequestScreen = props => {
 
         </View>
         </KeyboardAwareScrollView> 
+        </ImageBackground>
     );
 };
 
@@ -143,8 +144,11 @@ const styles = StyleSheet.create({
   container: {
     width:'100%',
     height:'100%',
-    backgroundColor: '#2c8ffa',
   
+},
+bImage: {
+  width:'100%',
+  height:'100%',
 },
   text: {
 
@@ -166,7 +170,6 @@ const styles = StyleSheet.create({
         width:'100%',
         height:'100%',
         flex: 1,
-        backgroundColor: '#2c8ffa',
         alignItems: 'center',
         justifyContent: 'center'
     },
