+++ screens/ManageRequestScreens.js
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from 'react';
-import { StyleSheet, Text, View, ScrollView,TextInput, TouchableOpacity,FlatList, SafeAreaView} from 'react-native';
+import { StyleSheet, Text, View, ImageBackground, TouchableOpacity,FlatList, SafeAreaView} from 'react-native';
 import {Card} from 'react-native-paper';
 import * as firebase from 'firebase';
 import deleteRequest from '../functions/deleteRequest';
@@ -62,6 +62,9 @@ const ManageRequestScreens = props => {
  
 
     return(
+
+        <ImageBackground source={require('../assets/BG.png')} style={styles.bImage}>
+
             
                 <View style={styles.screen} >
                 
@@ -76,6 +79,8 @@ const ManageRequestScreens = props => {
                 
                     
                 </View>
+
+                </ImageBackground>
                 
             
     );
@@ -91,17 +96,20 @@ headerTintColor:"white"
 };
 
 const styles = StyleSheet.create({
+    bImage: {
+        width:'100%',
+        height:'100%',
+    },
+    
     screen: {
         width:'100%',
         height:'100%',
         flex:1,
-        backgroundColor: '#2c8ffa',
         alignItems: 'center',
         justifyContent: 'center'
     },
     screen2: {
         width:410,
-        backgroundColor: '#2c8ffa',
         alignItems: 'center',
         justifyContent: 'center'
     },
