+++ screens/MapScreen.js
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from 'react';
-import { StyleSheet, Text, View, TextInput,TouchableOpacity } from 'react-native';
+import { StyleSheet, Text, View, TextInput,TouchableOpacity, ImageBackground } from 'react-native';
 import MapView, {Marker, Circle, Callout} from 'react-native-maps';
 import {Card} from 'react-native-paper';
 import * as Location from 'expo-location';
@@ -75,6 +75,8 @@ const MapScreen = props => {
 
 
     return (
+      <ImageBackground source={require('../assets/BG.png')} style={styles.bImage}>
+
     <View style={styles.screen} >
 
 <Card style={styles.cardView}>
@@ -138,6 +140,7 @@ const MapScreen = props => {
 
         
     </View>
+    </ImageBackground>
     );
 };
 
@@ -155,10 +158,13 @@ const styles = StyleSheet.create({
         width:'100%',
         height:'100%',
         flex: 1,
-        backgroundColor: '#2c8ffa',
         alignItems: 'center',
         justifyContent: 'center'
     },
+    bImage: {
+      width:'100%',
+      height:'100%',
+  },
     mapView: {
         width:'100%',
         height:'75%',
