+++ screens/PickUpAssistanceRequestScreen.js
@@ -1,5 +1,5 @@
 import React, { useState} from 'react';
-import { StyleSheet, Text, View,TextInput, TouchableOpacity,FlatList} from 'react-native';
+import { StyleSheet, Text, View,TextInput, TouchableOpacity,FlatList,ImageBackground} from 'react-native';
 import {Card} from 'react-native-paper';
 import * as firebase from 'firebase';
 import addVid from '../functions/addvidTwo';
@@ -84,6 +84,9 @@ const PickUpAssistanceRequestScreen = props => {
 
 
     return(
+
+        <ImageBackground source={require('../assets/BG.png')} style={styles.bImage}>
+
       <View style={styles.screen} >
       <Text style={styles.text}>Search or Add Request below:</Text>
       
@@ -116,6 +119,9 @@ const PickUpAssistanceRequestScreen = props => {
 
           
       </View>
+
+      </ImageBackground>
+        
       
     );
 };
@@ -135,15 +141,18 @@ const styles = StyleSheet.create({
       width:'100%',
       height:'100%',
       flex:1,
-      backgroundColor: '#2c8ffa',
       alignItems: 'center',
       justifyContent: 'center'
   },
   screen2: {
     width:410,
-    backgroundColor: '#2c8ffa',
     alignItems: 'center',
     justifyContent: 'center'
+},
+
+bImage: {
+    width:'100%',
+    height:'100%',
 },
 
   SubScreen: {
